Added some base classes for game logic/units.
Added 4 classes: WorldState, Controller, Army, Unit.

WorldState will be the "main" class, containing all the information for
the game as a whole.

Controller will be where we interface the player's input to the game
world / player's army.

Army is an interface for a collection of units.

Unit is a generic class represeting a unit of some kind.

Most functions are documented (not that there are many) and there's some
comments about what we might want to implement.